Move the hashtable mpmc tests under the data_structures folder
