+++ tests/data_structures/hashtable/mpmc/test-hashtable-mpmc-op-get.cpp
@@ -24,7 +24,7 @@
 #include "data_structures/hashtable/mcmp/hashtable_support_index.h"
 #include "data_structures/hashtable/mcmp/hashtable_op_get.h"
 
-#include "../../support.h"
+#include "../../../support.h"
 #include "fixtures-hashtable-mpmc.h"
 
 TEST_CASE("hashtable/hashtable_mcmp_op_get.c", "[hashtable][hashtable_op][hashtable_mcmp_op_get]") {
