+++ tests/data_structures/hashtable/mpmc/test-hashtable-mpmc-op-iter.cpp
@@ -24,7 +24,7 @@
 #include "data_structures/hashtable/mcmp/hashtable_support_index.h"
 #include "data_structures/hashtable/mcmp/hashtable_op_iter.h"
 
-#include "../../support.h"
+#include "../../../support.h"
 #include "fixtures-hashtable-mpmc.h"
 
 TEST_CASE("hashtable/hashtable_mcmp_op_iter.c", "[hashtable][hashtable_op][hashtable_mcmp_op_iter]") {
